refactor: Transfer keyboard data into the different folder.
